fix musicSelect / totalResult
譜面一覧をタイトルで取得するように変更 / 最終リザルトを表示

+++ src/UI/Scene/Play.hpp
@@ -30,8 +30,8 @@ namespace ui{
         Audio holdSound;
         EventTimer time;
         DecisionInfo decision;
-        float pointEachNote;
-        float point;
+        double pointEachNote;
+        double point;
         musicgame::judge_results_t results;
         Effect decisionEffect;
         Texture shine;
